+++ example/android/app/src/main/kotlin/in/juspay/hyper_webview_flutter_example/MainActivity.kt
@@ -1,10 +1,3 @@
-/*
- * Copyright (c) Juspay Technologies.
- *
- * This source code is licensed under the AGPL 3.0 license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
 package `in`.juspay.hyper_webview_flutter_example
 
 import io.flutter.embedding.android.FlutterActivity